+++ java/core/libjoynr/src/main/java/io/joynr/runtime/JoynrRuntimeImpl.java
@@ -50,6 +50,7 @@ import joynr.SubscriptionRequest;
 import joynr.SubscriptionStop;
 import joynr.system.DiscoveryProxy;
 import joynr.system.RoutingTypes.Address;
+import joynr.types.ProviderQos;
 
 import org.reflections.Reflections;
 import org.slf4j.Logger;
@@ -158,9 +159,26 @@ abstract public class JoynrRuntimeImpl implements JoynrRuntime {
         return proxyBuilderFactory.get(domain, interfaceClass);
     }
 
+    /**
+     * Registers a provider in the joynr framework
+     *
+     * @deprecated Will be removed by end of the year 2016. Use {@link io.joynr.runtime.JoynrRuntimeImpl#registerProvider(String, JoynrProvider, ProviderQos)} instead.
+     * @param domain
+     *            The domain the provider should be registered for. Has to be identical at the client to be able to find
+     *            the provider.
+     * @param provider
+     *            Instance of the provider implementation (has to extend a generated ...AbstractProvider).
+     * @return Returns a Future which can be used to check the registration status.
+     */
+    @Deprecated
     @Override
     public Future<Void> registerProvider(String domain, JoynrProvider provider) {
-        return capabilitiesRegistrar.registerProvider(domain, provider);
+        return capabilitiesRegistrar.registerProvider(domain, provider, provider.getProviderQos());
+    }
+
+    @Override
+    public Future<Void> registerProvider(String domain, JoynrProvider provider, ProviderQos providerQos) {
+        return capabilitiesRegistrar.registerProvider(domain, provider, providerQos);
     }
 
     @Override