+++ javascript/libjoynr-js/src/test/js/integration/TestEnd2EndCommProviderWorker.js
@@ -382,8 +382,7 @@ function initializeTest(provisioningSuffix, providedDomain) {
 
             providerQos.priority = Date.now();
             // register provider at the given providerDomain
-            libjoynrAsync.capabilities.registerCapability(
-                    "",
+            libjoynrAsync.registration.registerProvider(
                     providerDomain,
                     radioProvider,
                     providerQos).then(function() {
@@ -415,11 +414,10 @@ function startTest() {
             }, 10);
         }, mixedSubscriptionDelay);
         resolve(libjoynrAsync.participantIdStorage.getParticipantId(
-                "",
                 providerDomain,
                 radioProvider));
     });
 }
 function terminateTest() {
-    return libjoynrAsync.capabilities.unregisterCapability("", providerDomain, radioProvider);
+    return libjoynrAsync.registration.unregisterProvider(providerDomain, radioProvider);
 }